+++ CHANGELOG.md
@@ -3,12 +3,74 @@
 
 
 
+<a name="v1.6.0"></a>
+## [v1.6.0] - 2020-03-06
+
+- Added AWS S3 bucket public access block ([#18](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/issues/18))
+
+
+<a name="v1.5.0"></a>
+## [v1.5.0] - 2020-01-07
+
+- Updated CHANGELOG
+- Fix kms_master_key_id to conform with terraform resource ([#5](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/issues/5))
+
+
+<a name="v1.4.0"></a>
+## [v1.4.0] - 2019-11-22
+
+- Updated CHANGELOG
+- Fix for bucket policy count when value is not computed ([#12](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/issues/12))
+
+
+<a name="v1.3.0"></a>
+## [v1.3.0] - 2019-11-22
+
+- Updated CHANGELOG
+- Updated example to show bucket policy usage
+
+
+<a name="v1.2.0"></a>
+## [v1.2.0] - 2019-11-21
+
+- Updated CHANGELOG
+- Output bucket id from s3_bucket_policy to make sure that policy is present before it can be used
+
+
+<a name="v1.1.0"></a>
+## [v1.1.0] - 2019-11-20
+
+- Updated CHANGELOG
+- Added support for S3 bucket policy (incl. ELB logs delivery policy) ([#10](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/issues/10))
+
+
+<a name="v0.1.0"></a>
+## [v0.1.0] - 2019-08-29
+
+
+
+<a name="v1.0.0"></a>
+## [v1.0.0] - 2019-08-29
+
+- Updated CHANGELOG
+- Rewrite to match other modules, added all existing S3 features
+
+
 <a name="v0.0.1"></a>
 ## v0.0.1 - 2019-07-21
 
+- Updated CHANGELOG
 - Minor formatting, cleanups, readme
 - Initial release of this module ([#4](https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/issues/4))
 - Initial commit
 
 
-[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.6.0...HEAD
+[v1.6.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.5.0...v1.6.0
+[v1.5.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.4.0...v1.5.0
+[v1.4.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.3.0...v1.4.0
+[v1.3.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.2.0...v1.3.0
+[v1.2.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.1.0...v1.2.0
+[v1.1.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v0.1.0...v1.1.0
+[v0.1.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v1.0.0...v0.1.0
+[v1.0.0]: https://github.com/terraform-aws-modules/terraform-aws-s3-bucket/compare/v0.0.1...v1.0.0